requirements.txt
python-dotenv==0.16.0
pydantic
pytest